:necktie: Add useGetBlockchainApplicationByChainIdQuery for applications fetching
